Create pancake.cpp
creating a pancake class with swap functions and constructor to make it easier to make pancakes. Considering adding a vector function to the class in order to make it easier to swap the position of pancakes (not sure if that's already included in standard lib facilities vector or not). Also going to add additional variables to the constructor function as it will allow easier initial shuffling - will figure that out tomorrow morning.